feat: HealthSystem
HealthSystem that manages idol health;
Critical error: cannot get targetTransform when passing into function to ability!

+++ Assets/_Scripts/Scriptables/ScriptableIdolBase.cs
@@ -29,7 +29,7 @@ public struct Stats
 {
     public string idolName;
     public float health;
-    public float attackPower;
+    public int attackPower;
     public float attackRange;
     //public float attackCooldown;
     //public float specialAttackCooldown;
